all commit

+++ routes/listing.js
@@ -2,27 +2,14 @@ const express = require("express");
 const router = express.Router();
 const Listing = require('../models/listing.js');
 const wrapAsync = require('../utils/wrapasync');
-const ExpressError = require('../utils/ExpressError');
-const { listingSchema} = require('../schema.js');
-const { isLoggedIn } = require("../middleware");
 
-const validateListing = (req, res, next) => {
-    console.log(req.body);
-    let { error } = listingSchema.validate(req.body);
+const { isLoggedIn ,isOwner,validateListing } = require("../middleware");
 
-    if (error) {
-        let errMsg = error.details.map(el => el.message).join(",");
-        throw new ExpressError(400, errMsg);
-    }
 
-    next();
-};
+
 
 // index route
-router.get('/', wrapAsync(async (req, res) => {
-    const allListings = await Listing.find({});
-    res.render('listings/index.ejs', {  allListings });
-}));
+router.get('/', wrapAsync());
 
 // new route
 router.get('/new', isLoggedIn, (req, res) => {
@@ -32,8 +19,14 @@ router.get('/new', isLoggedIn, (req, res) => {
 // show route
 router.get('/:id', wrapAsync(async (req, res) => {
     const { id } = req.params;
-    const listing = await Listing.findById(id).populate("reviews");
+    const listing = await Listing.findById(id).populate("reviews").populate("owner");
+    if (!listing) {
+        req.flash("error", "Listing you request for does not exist");
+        return res.redirect("/listings");
+    }
+    console.log(listing);
     res.render('listings/show.ejs', { listing });
+
 }));
 
 // create route
@@ -44,6 +37,8 @@ router.post('/', isLoggedIn, (req, res, next) => {
     try{
     const newListing = new Listing(req.body.listing);
     console.log(newListing);
+    console.log(req.user);
+    newListing.owner = req.user._id; // Set the owner of the listing to the currently logged-in
     await newListing.save();
     console.log("saved Successfully");
     req.flash("success", "New Listing Created!");
@@ -56,23 +51,29 @@ router.post('/', isLoggedIn, (req, res, next) => {
 
 
 // edit route
-router.get('/:id/edit',isLoggedIn, wrapAsync(async (req, res) => {
+router.get('/:id/edit',isLoggedIn,isOwner, wrapAsync(async (req, res) => {
     const { id } = req.params;
     const listing = await Listing.findById(id);
     res.render('listings/edit.ejs', { listing });
 }));
 
 // update route
-router.put('/:id', isLoggedIn, validateListing, wrapAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isOwner, validateListing, wrapAsync(async (req, res) => {
     
     let { id } = req.params;
-    const listing = await Listing.findByIdAndUpdate(id, req.body.listing, { new: true });
-    res.redirect(`/listings/${listing._id}`);
+    await Listing.findByIdAndUpdate(id,{ ...req.body.listing });
+    req.flash("success", "Listing Updated Successfully!");
+    res.redirect(`/listings/${id}`);
 }));
 
 // delete route
-router.delete('/:id', isLoggedIn, wrapAsync(async (req, res) => {
+router.delete('/:id', isLoggedIn, isOwner, wrapAsync(async (req, res) => {
     const { id } = req.params;
+    let listing = await Listing.findById(id);
+    if (!listing.owner.equals(req.user._id)) {
+        req.flash("error", "You do not have permission to delete this listing.");
+        return res.redirect('/listings');
+    }
     await Listing.findByIdAndDelete(id);
     res.redirect('/listings');
 }));
